several UI improvoments

+++ app/src/main/java/me/cvhc/equationsolver/FunctionWrapper.java
@@ -105,7 +105,7 @@ public class FunctionWrapper implements XYSeries {
         return maxY;
     }
 
-    public double getNZero() {
+    public int getNZero() {
         return nZero;
     }
